refactor: remove firebase
