--- codegen.yml
@@ -1,18 +1,0 @@
-schema: http://localhost:8080/v1/graphql/
-documents:
-  - ./src/**/*.graphql
-overwrite: true
-generates:
-  ./src/generated/graphql.ts:
-    plugins:
-      - typescript
-      - typescript-operations
-      - typescript-react-apollo
-    config:
-      skipTypename: false
-      withHooks: true
-      withHOC: false
-      withComponent: false
-      scalars:
-        timestamptz: string
-        uuid: string
