--- firestore.rules
@@ -1,11 +1,0 @@
-rules_version = '2';
-service cloud.firestore {
-  match /databases/{database}/documents {
-     match /users/{userId} {
-      allow read: if request.auth != null;
-      allow create: if request.auth != null;
-      allow update: if request.auth.uid == userId;
-      allow delete: if request.auth.uid == userId;
-     }
-  }
-}
