--- functions/firebase.ts
@@ -1,21 +1,0 @@
-import { initializeApp } from 'firebase/app'
-import { getStorage } from 'firebase/storage'
-import { getAuth } from 'firebase/auth'
-import { getFirestore } from 'firebase/firestore/lite'
-
-const firebaseConfig = {
-    apiKey: process.env.FB_API_KEY,
-    authDomain: process.env.FB_AUTH_DOMAIN,
-    databaseURL: process.env.FB_DATABASE_URL,
-    projectId: process.env.FB_PROJECT_ID,
-    storageBucket: process.env.FB_STORAGE_BUCKET,
-    messagingSenderId: process.env.FB_MESSAGING_SENDER_ID,
-    appId: process.env.FB_APP_ID,
-}
-
-const app = initializeApp(firebaseConfig)
-const fbAuth = getAuth(app)
-const fbStorage = getStorage(app)
-const fbDb = getFirestore(app)
-
-export { fbAuth, fbStorage, fbDb }
