--- functions/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "functions",
-  "scripts": {
-    "build": "tsc",
-    "serve": "npm run build && firebase emulators:start --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log"
-  },
-  "engines": {
-    "node": "14"
-  },
-  "main": "lib/index.js",
-  "dependencies": {
-    "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.13.0",
-    "firebase-tools": "^9.1.0",
-    "nodemailer": "^6.4.17",
-    "react-redux": "^7.2.1"
-  },
-  "devDependencies": {
-    "firebase-functions-test": "^0.2.0",
-    "typescript": "^3.8.0"
-  },
-  "private": true
-}
