+++ src/components/common/icon/CloseIcon.tsx
