+++ src/components/common/icon/MinusCircle.tsx
