+++ src/components/layout/Main.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+const Main: React.FC = ({ children }) => {
+    return <main className="max-w-screen-md mx-auto mt-20 ">{children}</main>
+}
+
+export default Main
