--- src/components/layout/index.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const Layout: React.FC = ({ children }) => {
-    return <div className="flex-grow">{children}</div>
-}
-
-export default Layout
