+++ src/components/top/Form.tsx
@@ -5,7 +5,7 @@ import { uuidv4 } from '@firebase/util'
 import { EditTask } from '../../types/index'
 import { Project, Task, Category } from '@prisma/client'
 import { useCreateTasksByDate } from '@/hooks/task/useCreateTasksByDate'
-import CloseIcon from '@/components/common/CloseIcon'
+import CloseIcon from '@/components/common/icon/CloseIcon'
 import Button from '@/components/common/Button'
 import Row from './Row'
 import { SubmitHandler, useForm } from 'react-hook-form'
